+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 0.1.8 / 2024-02-14
+- HTMX and Hyperscript are now retrieved using NPM.
+- Fixes loading extensions from local/CDN and their paths. Thanks @agencyhub!
+
 # 0.1.7 / 2023-12-27
 - Bugfixes.